pyproject.toml
[tool.poetry]
name = "analysis-workspace"
version = "0.1.0"
description = "Workspace project for analysis"
authors = ["NEWROPE Co. Ltd. <data-science@newrope.biz>"]
packages = [
  {include = "project", from = "src"}
]

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
pandas = "^2.0.0"
scipy = "^1.8.0"
jupyterlab = "^3.4.2"
tqdm = "^4.64.0"
ipywidgets = "^8.0.6"
numpy = "^1.22.3"
matplotlib = "^3.7.1"
hydra-core = "^1.3.2"

[tool.poetry.group.dev.dependencies]
dvc = "^2.52.0"
dvc-s3 = "^2.21.0"
pytest = "^7.2.2"
tomlkit = "^0.11.7"
black = "^23.3.0"

[tool.poetry.scripts]
script = "project.cli:start"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
